.claude/adapters/tests/test_cmd_invoke_auth_type_wire_up.py
"""Cycle-110 sprint-2b2a — cmd_invoke → dispatch_filter wire-up integration.

Verifies the production dispatch path correctly:
1. Loads advisor_strategy config from .loa.config.yaml
2. Skips the filter when advisor_strategy.enabled=False (legacy behavior)
3. Runs filter + auto-mode when advisor_strategy.enabled=True
4. Threads auth_type_resolved + auth_type_selection_reason +
   auto_selection_inputs + auto_evaluation_timestamp into the MODELINV
   emitter kwargs

These tests use mocked adapters + mocked emit_model_invoke_complete to
avoid making real provider calls; they verify the WIRING, not the
substrate (which test_dispatch_filter.py covers separately).
"""

from __future__ import annotations

import json
import os
import sys
from pathlib import Path
from unittest.mock import MagicMock, patch

import pytest
import yaml

sys.path.insert(0, str(Path(__file__).resolve().parent.parent))


def _make_project_root(tmp_path, advisor_enabled, dispatch_preference="auto"):
    """Build a tmp project with model-config.yaml + .loa.config.yaml."""
    root = tmp_path
    (root / ".claude" / "defaults").mkdir(parents=True, exist_ok=True)
    model_config = {
        "providers": {
            "openai": {
                "type": "openai",
                "endpoint": "https://api.example.com/v1",
                "auth": "test-key",
                "models": {
                    "gpt-5.2": {
                        "capabilities": ["chat"],
                        "context_window": 128000,
                        "endpoint_family": "chat",
                        "auth_type": "http_api",
                        "dispatch_group": "openai-gpt",
                        "pricing": {
                            "input_per_mtok": 10000000,
                            "output_per_mtok": 30000000,
                        },
                    },
                },
            },
        },
        "aliases": {},
    }
    with (root / ".claude" / "defaults" / "model-config.yaml").open("w") as f:
        yaml.safe_dump(model_config, f, sort_keys=False)

    loa_config = {
        "advisor_strategy": {
            "schema_version": 2,
            "enabled": advisor_enabled,
            "dispatch_preference": dispatch_preference,
            "defaults": {
                "planning": "advisor",
                "review": "advisor",
                "audit": "advisor",
                "implementation": "advisor",
            },
            "tier_aliases": {
                "advisor": {"openai": "gpt-5.2"},
                "executor": {"openai": "gpt-5.2"},
            },
        },
    }
    with (root / ".loa.config.yaml").open("w") as f:
        yaml.safe_dump(loa_config, f, sort_keys=False)
    return str(root)


class TestAdvisorStrategyDisabledLegacy:
    """When advisor_strategy is absent / disabled, the wire-up MUST be a
    no-op (preserves pre-cycle-110 behavior exactly).

    This test class exercises the canonical AdvisorStrategyConfig.disabled_legacy()
    factory — the actual cmd_invoke gating uses `if _adv_cfg.enabled:` so a
    disabled config means the dispatch_filter / run_auto_mode pipeline is
    skipped entirely.
    """

    def test_disabled_legacy_has_enabled_false(self):
        from loa_cheval.config.advisor_strategy import AdvisorStrategyConfig
        cfg = AdvisorStrategyConfig.disabled_legacy()
        assert cfg.enabled is False

    def test_disabled_legacy_dispatch_preference_default_auto(self):
        from loa_cheval.config.advisor_strategy import AdvisorStrategyConfig
        cfg = AdvisorStrategyConfig.disabled_legacy()
        # The disabled-legacy factory carries default values — but enabled=False
        # gates the wire-up so these values are never read in production.
        assert cfg.dispatch_preference == "auto"


class TestDispatchFilterIntegration:
    """Smoke-tests for the dispatch_filter + run_auto_mode pipeline that
    cmd_invoke calls. The integration is verified more fully in
    test_dispatch_filter.py; here we pin the cmd_invoke contract surface."""

    def test_auto_mode_cold_start_with_headless_chain_resolves_headless(self):
        """When auto-mode runs on a chain with a headless entry and no
        warm stats, the resolution defaults to headless."""
        from loa_cheval.routing.dispatch_filter import (
            run_auto_mode, SELECTION_REASON_AUTO_DEFAULT,
        )
        from loa_cheval.routing.types import ResolvedChain, ResolvedEntry

        entries = (
            ResolvedEntry(
                provider="anthropic", model_id="claude-headless",
                adapter_kind="cli", capabilities=frozenset(["chat"]),
                auth_type="headless", dispatch_group="anthropic-claude",
            ),
            ResolvedEntry(
                provider="anthropic", model_id="claude-opus-4-7",
                adapter_kind="http", capabilities=frozenset(["chat"]),
                auth_type="http_api", dispatch_group="anthropic-claude",
            ),
        )
        chain = ResolvedChain(
            primary_alias="claude-headless",
            entries=entries,
            headless_mode="prefer-api",
            headless_mode_source="default",
        )

        ar = run_auto_mode(chain, stats={}, advisor_config={}, now=1.0)
        assert ar.selected_auth_type == "headless"
        assert ar.reason == SELECTION_REASON_AUTO_DEFAULT
        assert ar.evaluation_timestamp == 1.0

    def test_auto_resolution_envelope_inputs_only_for_band_comparison(self):
        """as_selection_inputs() is only meaningful when reason=auto-band-comparison.
        For cold-start paths, the inputs dict is technically present but
        should be skipped by the emitter wire-up (sprint-2b2a wire-up logic)."""
        from loa_cheval.routing.dispatch_filter import (
            AutoResolution, SELECTION_REASON_AUTO_BAND,
            SELECTION_REASON_AUTO_DEFAULT,
        )

        warm = AutoResolution(
            selected_auth_type="headless",
            reason=SELECTION_REASON_AUTO_BAND,
            evaluation_timestamp=1.0,
            sample_n_per_bucket={"headless": 100},
            band_per_bucket={"headless": "green"},
            success_rate_per_bucket={"headless": 0.95},
        )
        # Warm path → inputs populated.
        assert warm.as_selection_inputs()["sample_n_per_bucket"] == {"headless": 100}

        cold = AutoResolution(
            selected_auth_type="headless",
            reason=SELECTION_REASON_AUTO_DEFAULT,
            evaluation_timestamp=1.0,
        )
        # Cold-start path → inputs are empty (caller checks reason and skips).
        assert cold.as_selection_inputs() == {
            "sample_n_per_bucket": {},
            "band_per_bucket": {},
            "success_rate_per_bucket": {},
        }


class TestEmitterAcceptsCycle110Kwargs:
    """The MODELINV emitter MUST accept the cycle-110 v1.4 kwargs cleanly.
    Wire-up in cheval.cmd_invoke depends on this contract."""

    def test_emit_with_all_cycle_110_fields(self):
        from loa_cheval.audit.modelinv import emit_model_invoke_complete
        # When LOA_MODELINV_AUDIT_DISABLE=1, emitter no-ops post-redaction.
        with patch.dict(os.environ, {"LOA_MODELINV_AUDIT_DISABLE": "1"}):
            emit_model_invoke_complete(
                models_requested=["anthropic:claude-headless"],
                models_succeeded=["anthropic:claude-headless"],
                models_failed=[],
                operator_visible_warn=False,
                auth_type_resolved="headless",
                auth_type_selection_reason="auto-cold-start-default-headless",
                auto_selection_inputs={
                    "sample_n_per_bucket": {"headless": 0},
                    "band_per_bucket": {},
                    "success_rate_per_bucket": {},
                },
                auto_evaluation_timestamp=1715789012.345,
            )

    def test_emit_rejects_invalid_auth_type_resolved(self):
        from loa_cheval.audit.modelinv import emit_model_invoke_complete
        with patch.dict(os.environ, {"LOA_MODELINV_AUDIT_DISABLE": "1"}):
            with pytest.raises(ValueError, match="auth_type_resolved"):
                emit_model_invoke_complete(
                    models_requested=["anthropic:claude-headless"],
                    models_succeeded=["anthropic:claude-headless"],
                    models_failed=[],
                    operator_visible_warn=False,
                    auth_type_resolved="subscription",  # not in enum
                )

    def test_emit_legacy_call_without_cycle_110_kwargs_still_works(self):
        """Backward-compat: callers that DON'T pass the new kwargs produce
        a v1.3-shape envelope (additive evolution preserved)."""
        from loa_cheval.audit.modelinv import emit_model_invoke_complete
        with patch.dict(os.environ, {"LOA_MODELINV_AUDIT_DISABLE": "1"}):
            emit_model_invoke_complete(
                models_requested=["anthropic:claude-opus-4-7"],
                models_succeeded=["anthropic:claude-opus-4-7"],
                models_failed=[],
                operator_visible_warn=False,
            )
